pyproject.toml
[project]
name = "final-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faker>=37.8.0",
    "kafka-python>=2.2.15",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
]
